Moved User class
