+++ src/main/java/Application/Database/User/UserRepository.java
@@ -1,10 +1,9 @@
 package Application.Database.User;
 
-import Application.Entities.User.User;
+import Application.Entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 
 public interface UserRepository extends JpaRepository<User, Long>, CustomUsers {
