+++ src/main/java/Application/Entities/Role.java
@@ -1,6 +1,5 @@
 package Application.Entities;
 
-import Application.Entities.User.User;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.security.core.GrantedAuthority;
