+++ src/main/java/Application/Entities/User.java
@@ -1,14 +1,11 @@
-package Application.Entities.User;
+package Application.Entities;
 
-import Application.Entities.Role;
 import lombok.*;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import javax.persistence.*;
 import javax.validation.constraints.Size;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.Collection;
 import java.util.Date;
 import java.util.Set;
